add readme.md

+++ src/main/java/com/recycleme/actionListener/jenis/JenisHapus.java
@@ -4,6 +4,7 @@ import com.recycleme.dao.JenisDao;
 import com.recycleme.frame.jenis.JenisFrame;
 import com.recycleme.model.jenis.Jenis;
 
+import javax.swing.*;
 import java.awt.event.*;
 
 public class JenisHapus implements ActionListener {
@@ -17,8 +18,20 @@ public class JenisHapus implements ActionListener {
 
     public void actionPerformed(ActionEvent e) {
         if(e.getSource() == jenisFrame.getButtonDeleteJenis()) {
-            int id = jenisFrame.getSelectedJenisId();
             int row = jenisFrame.getSelectedJenisRow();
+            if(row == -1) {
+                jenisFrame.showErrorMessage("Pilih kategori terlebih dahulu!");
+                return;
+            }
+            int id = jenisFrame.getSelectedJenisId();
+            int confirm = JOptionPane.showConfirmDialog(jenisFrame, "Apakah anda yakin ingin menghapus jenis " +
+                            "ini?",
+                    "Konfirmasi", JOptionPane.YES_NO_OPTION);
+
+            if(confirm == JOptionPane.NO_OPTION) {
+                jenisFrame.showInfoMessage("Proses dibatalkan!");
+                return;
+            }
             if(jenisDao.delete(id) == 0) {
                 jenisFrame.showErrorMessage("Data gagal dihapus!");
                 return;
